+++ src/python/phenix_apps/apps/ignition/__init__.py
@@ -1,0 +1,3 @@
+from .app import Ignition, IgnitionHostConfig, RtuDeviceConfig
+
+__all__ = ["Ignition", "IgnitionHostConfig", "RtuDeviceConfig"]