+++ src/python/phenix_apps/apps/ignition/__main__.py
@@ -1,0 +1,9 @@
+from .app import Ignition
+
+
+def main():
+    Ignition.main("ignition")
+
+
+if __name__ == "__main__":
+    main()